update routing, split addGoal to setup, view to dashboard
