+++ src/components/Goals/EditGoal/EditGoal.tsx
@@ -12,7 +12,7 @@ import {
   Typography
 } from '@material-ui/core';
 import useStyles from './styles';
-import GoalModel from '../../../models/goal';
+import GoalModel from '../../../types/goal';
 import { useDispatch, useSelector } from 'react-redux';
 import { ERROR } from '../../../constants/actionTypes';
 import Alert from '../../../helpers/Alert';
