+++ src/components/Goals/Goal.tsx
@@ -14,7 +14,7 @@ import {
 } from '@material-ui/core';
 import { useHistory } from 'react-router-dom';
 import useStyles from './styles';
-import GoalModel from '../../models/goal';
+import GoalModel from '../../types/goal';
 import DeleteIcon from '@material-ui/icons/Delete';
 import EditRoundedIcon from '@material-ui/icons/EditRounded';
 import { deleteGoal } from '../../store/actions/goals';
@@ -22,10 +22,11 @@ import { useDispatch } from 'react-redux';
 
 interface props {
   goal: GoalModel;
+  setupView: Boolean;
 }
 
 const Goal = (props: props) => {
-  const { goal } = props;
+  const { goal, setupView } = props;
   const {
     title,
     description,
@@ -65,15 +66,17 @@ const Goal = (props: props) => {
           >
             Description: {description}
           </Typography>
-          <Typography
-            variant="body2"
-            color="textSecondary"
-            component="p"
-            gutterBottom
-            noWrap
-          >
-            Cadence: {cadence}
-          </Typography>
+          {setupView &&
+            <Typography
+              variant="body2"
+              color="textSecondary"
+              component="p"
+              gutterBottom
+              noWrap
+            >
+              Cadence: {cadence}
+            </Typography>
+          }
           <FormControlLabel
             label="Complete?"
             control={
@@ -85,26 +88,28 @@ const Goal = (props: props) => {
             }
           />
         </CardContent>
-        <CardActions>
-          {loggedUser && Object.keys(loggedUser).length !== 0 && userID === goalUserID &&
-            <>
-              <IconButton
-                title="Edit goal"
-                aria-label="edit goal"
-                onClick={() => history.push(`/editGoal/${props.goal._id}`)}
-              >
-                <EditRoundedIcon fontSize="large" color="secondary" />
-              </IconButton>
-              <IconButton
-                title="Delete goal"
-                aria-label="delete goal"
-                onClick={() => dispatch(deleteGoal(props.goal._id, history))}
-              >
-                <DeleteIcon fontSize="large" color="secondary" />
-              </IconButton>
-            </>
-          }
-        </CardActions>
+        {setupView &&
+          <CardActions>
+            {loggedUser && Object.keys(loggedUser).length !== 0 && userID === goalUserID &&
+              <>
+                <IconButton
+                  title="Edit goal"
+                  aria-label="edit goal"
+                  onClick={() => history.push(`/editGoal/${props.goal._id}`)}
+                >
+                  <EditRoundedIcon fontSize="large" color="secondary" />
+                </IconButton>
+                <IconButton
+                  title="Delete goal"
+                  aria-label="delete goal"
+                  onClick={() => dispatch(deleteGoal(props.goal._id, history))}
+                >
+                  <DeleteIcon fontSize="large" color="secondary" />
+                </IconButton>
+              </>
+            }
+          </CardActions>
+        }
       </Card>
     </Grid>
   );
