+++ src/components/Navbar/Navbar.tsx
@@ -2,13 +2,18 @@ import React, { useEffect, useState } from "react";
 import useStyles from './styles';
 import { 
     AppBar,
-    // Box,
+    Box,
     Button,
-    // MenuItem,
+    Grid,
+    MenuItem,
     Toolbar,
     Typography
 } from "@material-ui/core";
-import { Link, useHistory, useLocation } from 'react-router-dom';
+import { 
+    Link,
+    useHistory,
+    useLocation 
+} from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 import * as actionType from '../../constants/actionTypes';
 import decode from 'jwt-decode';
@@ -18,9 +23,11 @@ const Navbar = () => {
     const classes = useStyles();
     const history = useHistory();
     const location = useLocation();
-    const handleLogin = () => history.push('/loginUser');
     const profile = localStorage.getItem('profile')!;
+
     const [user, setUser] = useState(JSON.parse(profile));
+
+    const handleLogin = () => history.push('/loginUser');
     const dispatch = useDispatch();
 
     const logout = () => {
@@ -28,6 +35,8 @@ const Navbar = () => {
         history.push('/');
         setUser(null);
     };
+
+    const homePath = (user ? "/dashboard" : "/");
 
     useEffect(() => {
         const token = user?.token;
@@ -38,45 +47,70 @@ const Navbar = () => {
         }
 
         setUser(JSON.parse(profile));
+
         // eslint-disable-next-line
     }, [location]);
 
     return (
-        <div className={classes.root} style={{ position: "absolute", left: "0", top: "0" }}>
+        <Grid container>
             <AppBar color="primary">
                 <Toolbar>
-                    <Typography variant="h6" className={classes.title} align="left">
+                    <Typography 
+                        variant="h6" 
+                        className={classes.title} 
+                        align="left"
+                    >
                         <Link
-                            to="/"
+                            to={homePath}
                             style={{ textDecoration: "none", color: "#fff" }}
                         >
                             Goal Tracker
                         </Link>
                     </Typography>
                     {user?.result ?
-                        <div style={{ display: "flex" }}>
-                            <Avatar
-                                className={classes.purple}
-                                alt={user?.result.name}
-                                src={user?.result.imageUrl}
-                            >
-                                {user?.result.name.charAt(0)}
-                            </Avatar>
-                            <Typography
-                                className={classes.userName}
-                                variant="h6"
-                            >
-                                {user?.result.name}
-                            </Typography>
-                            <Button
-                                variant="contained"
-                                style={{ marginLeft: "20px" }}
-                                color="secondary"
-                                onClick={logout}
-                            >
-                                Logout
-                            </Button>
-                        </div>
+                        <>
+                            <Grid item xs={4} style={{ display: 'flex' }}>
+                                <Link to="/dashboard" style={{ textDecoration: 'none', padding: 10 }}>
+                                    <Button
+                                        color="secondary"
+                                        variant="outlined"
+                                    >
+                                        Dashboard
+                                    </Button>
+                                </Link>
+                                <Link to="/setup" style={{ textDecoration: 'none', padding: 10 }}>
+                                    <Button
+                                        color="secondary"
+                                        variant="outlined"
+                                    >
+                                        Setup    
+                                    </Button>
+                                </Link>
+                            </Grid>
+                            <div style={{ display: "flex" }}>
+                                <Avatar
+                                    className={classes.purple}
+                                    alt={user?.result.name}
+                                    src={user?.result.imageUrl}
+                                >
+                                    {user?.result.name.charAt(0)}
+                                </Avatar>
+                                <Typography
+                                    className={classes.userName}
+                                    variant="h6"
+                                >
+                                    {user?.result.name}
+                                </Typography>
+                                <Button
+                                    variant="contained"
+                                    style={{ marginLeft: "20px" }}
+                                    color="secondary"
+                                    onClick={logout}
+                                >
+                                    Logout
+                                </Button>
+                            </div>
+                        </>
                         :
                         <Button
                             color="primary"
@@ -88,7 +122,7 @@ const Navbar = () => {
                     }
                 </Toolbar>
             </AppBar>
-        </div>
+        </Grid>
     );
 };
