+++ src/components/Router/Router.tsx
@@ -6,8 +6,8 @@ import RegisterUser from '../Auth/RegisterUser';
 import AddGoal from '../Goals/AddGoal/AddGoal';
 import EditGoal from '../Goals/EditGoal/EditGoal';
 import Dashboard from '../Dashboard/Dashboard';
-import Goals from '../Goals/Goals';
 import LandingPage from '../LandingPage/LandingPage'
+import SetupPage from '../SetupPage/SetupPage'
 
 const Router = () => {
 
@@ -15,9 +15,9 @@ const Router = () => {
     <Switch>
       <Route path="/" exact component={LandingPage} />
       <Route path="/dashboard" exact component={Dashboard} />
+      <Route path="/setup" exact component={SetupPage} />
       <Route path="/addGoal" exact component={AddGoal} />
       <Route path="/editGoal/:id" exact component={EditGoal} />
-      <Route path="/goals" exact component={Goals} />
       <Route path="/loginUser" exact component={LoginUser} />
       <Route path="/register" exact component={RegisterUser} />
       <Route path="/unauthorized" exact component={Unauthorized} />
