+++ src/components/SetupPage/SetupPage.tsx
@@ -15,13 +15,13 @@ import { useHistory } from 'react-router';
 import { 
   getUserGoals 
 } from '../../store/actions/goals';
-import GoalModel from '../../models/goal';
-import Goal from './Goal';
+import GoalModel from '../../types/goal';
+import Goal from '../../components/Goals/Goal';
 import { DELETE_SUCCESSFUL, UPDATE_SUCCESSFUL } from '../../constants/actionTypes';
 import Alert from '../../helpers/Alert';
 import { displayGoalOnCadence } from '../../helpers/cadence'
 
-const Goals = () => {
+const SetupPage = () => {
   const {
     goals,
     isLoading,
@@ -85,12 +85,13 @@ const Goals = () => {
 
   return (
     <Container style={{ marginTop: "100px" }}>
-      <Typography variant="h1">Today's Goals</Typography>
       {user?.result &&
         <Button
           variant="contained"
           color="primary"
           onClick={() => history.push("/addGoal")}
+          fullWidth
+          style={{ marginBottom: 20 }}
         >
           Add goal
         </Button>
@@ -104,7 +105,7 @@ const Goals = () => {
           <Grid container spacing={3}>
             {
               goals?.map((goal: GoalModel, index: number) => (
-                <Goal goal={goal} key={index} />
+                <Goal goal={goal} key={index} setupView={true} />
               ))
             }
           </Grid>
@@ -138,4 +139,4 @@ const Goals = () => {
   );
 };
 
-export default Goals;
+export default SetupPage;
